# spring-boot-springfox/src/main/resources/static/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Spring-boot-Springfox Example",
    "version": "2.0",
    "title": "Document Api",
    "contact": {
      "name": "lance",
      "url": "https://github.com/leelance",
      "email": "81222045@qq.com"
    },
    "license": {
      "name": "Apache License Version 2.0"
    }
  },
  "host": "localhost",
  "basePath": "/",
  "tags": [
    {
      "name": "home-controller",
      "description": "Home Controller"
    },
    {
      "name": "index-controller",
      "description": "Index Controller"
    },
    {
      "name": "fox-controller",
      "description": "Fox Controller"
    }
  ],
  "paths": {
    "/api/v1.0/delete/{id}": {
      "get": {
        "tags": [
          "index-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1.0/detail/id_{id}": {
      "get": {
        "tags": [
          "index-controller"
        ],
        "summary": "detail",
        "operationId": "detailUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1.0/home/delete/{id}": {
      "get": {
        "tags": [
          "home-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "fox-controller"
        ],
        "summary": "remove",
        "operationId": "removeUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1.0/home/findAll": {
      "get": {
        "tags": [
          "fox-controller"
        ],
        "summary": "findAll",
        "operationId": "findAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "params",
            "in": "query",
            "description": "params",
            "required": false,
            "type": "object"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1.0/home/list": {
      "get": {
        "tags": [
          "home-controller"
        ],
        "summary": "list",
        "operationId": "listUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1.0/home/save": {
      "post": {
        "tags": [
          "home-controller"
        ],
        "summary": "save",
        "operationId": "saveUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "text",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1.0/home/view": {
      "get": {
        "tags": [
          "fox-controller"
        ],
        "summary": "view",
        "operationId": "viewUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": true,
            "type": "string"
          },
          {
            "name": "address",
            "in": "query",
            "description": "address",
            "required": true,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1.0/list": {
      "get": {
        "tags": [
          "index-controller"
        ],
        "summary": "list",
        "operationId": "listUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1.0/save": {
      "post": {
        "tags": [
          "index-controller"
        ],
        "summary": "save",
        "operationId": "saveUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "text",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}